[core] Split RenderStyle from Style
