+++ src/mbgl/annotation/render_annotation_source.hpp
@@ -29,7 +29,7 @@ public:
     std::unordered_map<std::string, std::vector<Feature>>
     queryRenderedFeatures(const ScreenLineString& geometry,
                           const TransformState& transformState,
-                          const style::Style& style,
+                          const RenderStyle& style,
                           const RenderedQueryOptions& options) const final;
 
     std::vector<Feature>
