+++ src/mbgl/renderer/painter.cpp
@@ -2,6 +2,7 @@
 #include <mbgl/renderer/paint_parameters.hpp>
 #include <mbgl/renderer/render_tile.hpp>
 #include <mbgl/renderer/render_source.hpp>
+#include <mbgl/renderer/render_style.hpp>
 
 #include <mbgl/style/source.hpp>
 #include <mbgl/style/source_impl.hpp>
@@ -11,7 +12,6 @@
 #include <mbgl/util/logging.hpp>
 #include <mbgl/gl/debugging.hpp>
 
-#include <mbgl/style/style.hpp>
 #include <mbgl/style/layer_impl.hpp>
 #include <mbgl/style/layers/custom_layer_impl.hpp>
 
@@ -126,7 +126,7 @@ void Painter::cleanup() {
     context.performCleanup();
 }
 
-void Painter::render(const Style& style, const FrameData& frame_, View& view) {
+void Painter::render(RenderStyle& style, const FrameData& frame_, View& view) {
     frame = frame_;
     if (frame.contextMode == GLContextMode::Shared) {
         context.setDirtyState();
