+++ src/mbgl/renderer/sources/render_image_source.hpp
@@ -43,7 +43,7 @@ public:
     std::unordered_map<std::string, std::vector<Feature>>
     queryRenderedFeatures(const ScreenLineString& geometry,
                           const TransformState& transformState,
-                          const style::Style& style,
+                          const RenderStyle& style,
                           const RenderedQueryOptions& options) const final;
 
     std::vector<Feature> querySourceFeatures(const SourceQueryOptions&) const final;
