+++ src/mbgl/renderer/sources/render_raster_source.cpp
@@ -71,7 +71,7 @@ std::map<UnwrappedTileID, RenderTile>& RenderRasterSource::getRenderTiles() {
 std::unordered_map<std::string, std::vector<Feature>>
 RenderRasterSource::queryRenderedFeatures(const ScreenLineString&,
                                           const TransformState&,
-                                          const style::Style&,
+                                          const RenderStyle&,
                                           const RenderedQueryOptions&) const {
     return {};
 }
