+++ src/mbgl/renderer/update_parameters.hpp
@@ -13,14 +13,22 @@ class AnnotationManager;
 class UpdateParameters {
 public:
     const MapMode mode;
-    const Update updateFlags;
     const float pixelRatio;
     const MapDebugOptions debugOptions;
     const TimePoint timePoint;
-    const TransformState& transformState;
+
+    const std::string glyphURL;
+    const bool spriteLoaded;
+    const style::TransitionOptions transitionOptions;
+    const Immutable<style::Light::Impl> light;
+    const std::vector<Immutable<style::Image::Impl>> images;
+    const std::vector<Immutable<style::Source::Impl>> sources;
+    const std::vector<Immutable<style::Layer::Impl>> layers;
+
     Scheduler& scheduler;
     FileSource& fileSource;
     AnnotationManager& annotationManager;
+    const TransformState& transformState;
 };
 
 } // namespace mbgl
